codex: address PR review feedback (#20718)
