+++ codex-rs/app-server-daemon/Cargo.toml
@@ -13,19 +13,22 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-codex-app-server = { workspace = true }
 codex-app-server-protocol = { workspace = true }
+codex-app-server-transport = { workspace = true }
 codex-core = { workspace = true }
 codex-uds = { workspace = true }
 futures = { workspace = true }
 libc = { workspace = true }
+reqwest = { workspace = true, features = ["rustls-tls"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = [
     "fs",
+    "io-util",
     "macros",
     "process",
     "rt-multi-thread",
+    "signal",
     "time",
 ] }
 tokio-tungstenite = { workspace = true }
